add navigationView change design

+++ app/src/main/java/ru/zkdev/parking/viewModels/HistoryVM.java
@@ -18,6 +18,7 @@ public class HistoryVM extends ViewModel {
   }
 
   public void geleteById(int id) {
+    if (id <= 0)return;
     repository.deleteById(id);
   }
